Move logic of adjust pmem threshold to zmalloc module
- remove adjust ratio in RDB/AOF mechanism
- remove memory-ratio-check-period parameter
