+++ redis.conf
@@ -1120,8 +1120,5 @@ dynamic-threshold-min 24
 # Maximum value of dynamic threshold
 dynamic-threshold-max 10000
 
-# DRAM/PMEM ratio period measured in miliseconds
-memory-ratio-check-period 100
-
 # Keep hashtable structure always on DRAM
 hashtable-on-dram yes
