+++ src/server.c
@@ -1032,13 +1032,6 @@ void databasesCron(void) {
     if (server.active_expire_enabled && server.masterhost == NULL)
         activeExpireCycle(ACTIVE_EXPIRE_CYCLE_SLOW);
 
-    /* Adjust PMEM threshold. */
-    if (server.memory_alloc_policy == MEM_POLICY_RATIO) {
-        run_with_period(server.ratio_check_period) {
-            adjustPmemThresholdCycle();
-        }
-    }
-
     /* Perform hash tables rehashing if needed, but only if there are no
      * other processes saving the DB on disk. Otherwise rehashing is bad
      * as will cause a lot of copy-on-write of memory pages. */
@@ -1526,7 +1519,6 @@ void initServerConfig(void) {
     server.maxclients = CONFIG_DEFAULT_MAX_CLIENTS;
     server.hashtable_on_dram = 1;
     server.memory_alloc_policy = MEM_POLICY_ONLY_DRAM;
-    server.ratio_check_period = 100;
     server.initial_dynamic_threshold = 64;
     server.dynamic_threshold_min = 24;
     server.dynamic_threshold_max = 10000;
@@ -1884,6 +1876,28 @@ void resetServerStats(void) {
     server.stat_net_input_bytes = 0;
     server.stat_net_output_bytes = 0;
     server.aof_delayed_fsync = 0;
+}
+
+/* Initialize the pmem threshold. */
+static void pmemThresholdInit(void) {
+    switch(server.memory_alloc_policy) {
+        case MEM_POLICY_ONLY_DRAM:
+            zmalloc_set_threshold(UINT_MAX);
+            break;
+        case MEM_POLICY_ONLY_PMEM:
+            zmalloc_set_threshold(0U);
+            break;
+        case MEM_POLICY_THRESHOLD:
+            zmalloc_set_threshold(server.static_threshold);
+            break;
+        case MEM_POLICY_RATIO:
+            zmalloc_set_threshold(server.initial_dynamic_threshold);
+            zmalloc_set_ratio(server.target_pmem_dram_ratio);
+            zmalloc_set_threshold_range(server.dynamic_threshold_min, server.dynamic_threshold_max);
+            break;
+        default:
+            serverAssert(NULL);
+    }
 }
 
 void initServer(void) {
